fix: change chat_id type to fit telegram's chat ids

+++ get_updates_response.hpp
@@ -7,7 +7,7 @@
 struct chat {
   chat() = default;
 
-  int id = {};
+  double id = {};
 };
 
 struct message {
